+++ odbcsv/Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.4
+
+* Updated dependencies
+
 ## 0.2.3
 
 * Updated dependencies